# public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "S5ARC. API Documentation",
    "description": "API documentation for S5ARC.",
    "version": "1.0.0",
    "contact": {
      "email": "s5sajidyt+s5arc_dev@gmail.com",
      "name": "S5SAJID (SajidUllah K.)",
      "url": "https://s5sajid.github.io"
    }
  },
  "paths": {
    "/api/v1/customers/": {
      "get": {
        "operationId": "listCustomers",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCustomers200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCustomers401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCustomers403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCustomers404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCustomers500ResponseBody"
                }
              }
            }
          }
        },
        "tags": ["Customers"]
      }
    },
    "/api/v1/customers/{id}/": {
      "get": {
        "operationId": "getCustomer",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCustomer200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCustomer401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCustomer403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCustomer404ResponseBody2"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCustomer500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Customers"]
      }
    },
    "/api/v1/integrations/": {
      "get": {
        "operationId": "listIntegrations",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListIntegrations200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListIntegrations401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListIntegrations403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListIntegrations404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListIntegrations500ResponseBody"
                }
              }
            }
          }
        },
        "tags": ["Integrations"]
      }
    },
    "/api/v1/integrations/{id}/": {
      "get": {
        "operationId": "getIntegration",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIntegration200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIntegration401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIntegration403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIntegration404ResponseBody2"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIntegration500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Integrations"]
      },
      "post": {
        "operationId": "installIntegration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallIntegrationRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for status 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallIntegration201ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallIntegration401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallIntegration403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallIntegration404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallIntegration500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Integrations"]
      },
      "patch": {
        "operationId": "updateIntegration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateIntegration200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateIntegration401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateIntegration403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateIntegration404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateIntegration500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Integrations"]
      },
      "delete": {
        "operationId": "uninstallIntegration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UninstallIntegrationRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UninstallIntegration200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UninstallIntegration401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UninstallIntegration403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UninstallIntegration404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UninstallIntegration500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Integrations"]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetCustomer200ResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "name": { "type": "string", "minLength": 2, "maxLength": 100 },
              "email": {
                "type": "string",
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "phone": { "type": "string", "minLength": 7, "maxLength": 15 },
              "orders": { "type": "number", "minimum": 0 },
              "total_spent": { "type": "number", "minimum": 0 }
            },
            "required": ["name", "email"],
            "additionalProperties": false
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "GetCustomer401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetCustomer403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "GetCustomer404ResponseBody2": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetCustomer500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetIntegration200ResponseBody": {
        "type": "object",
        "properties": {
          "definition": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": { "type": "string", "maxLength": 255 },
              "slug": { "type": "string", "maxLength": 255 },
              "description": { "nullable": true, "type": "string" },
              "logo_url": {
                "nullable": true,
                "type": "string",
                "maxLength": 500
              },
              "target_endpoint_url": { "type": "string", "maxLength": 500 },
              "subscribed_events": {
                "anyOf": [
                  {
                    "anyOf": [
                      { "type": "string" },
                      { "type": "number" },
                      { "type": "boolean" },
                      { "type": "string", "nullable": true, "enum": [null] }
                    ]
                  },
                  { "type": "object", "additionalProperties": {} },
                  { "type": "array", "items": {} }
                ]
              },
              "config_schema": {
                "anyOf": [
                  {
                    "anyOf": [
                      { "type": "string" },
                      { "type": "number" },
                      { "type": "boolean" },
                      { "type": "string", "nullable": true, "enum": [null] }
                    ]
                  },
                  { "type": "object", "additionalProperties": {} },
                  { "type": "array", "items": {} }
                ]
              },
              "is_global": { "type": "boolean" },
              "created_at": {},
              "updated_at": {}
            },
            "required": [
              "id",
              "name",
              "slug",
              "description",
              "logo_url",
              "target_endpoint_url",
              "subscribed_events",
              "config_schema",
              "is_global",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false
          },
          "installation": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "integration_id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "is_enabled": { "type": "boolean" },
              "config": {
                "anyOf": [
                  {
                    "anyOf": [
                      { "type": "string" },
                      { "type": "number" },
                      { "type": "boolean" },
                      { "type": "string", "nullable": true, "enum": [null] }
                    ]
                  },
                  { "type": "object", "additionalProperties": {} },
                  { "type": "array", "items": {} }
                ]
              },
              "created_at": {},
              "updated_at": {}
            },
            "required": [
              "id",
              "integration_id",
              "is_enabled",
              "config",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false
          }
        },
        "required": ["definition", "installation"],
        "additionalProperties": false
      },
      "GetIntegration401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetIntegration403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "GetIntegration404ResponseBody2": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetIntegration500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "InstallIntegrationRequestBody": {
        "type": "object",
        "properties": {
          "config": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          }
        },
        "required": ["config"]
      },
      "ListCustomers200ResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "minLength": 2, "maxLength": 100 },
                "email": {
                  "type": "string",
                  "format": "email",
                  "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                },
                "phone": { "type": "string", "minLength": 7, "maxLength": 15 },
                "orders": { "type": "number", "minimum": 0 },
                "total_spent": { "type": "number", "minimum": 0 }
              },
              "required": ["name", "email"],
              "additionalProperties": false
            }
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "ListCustomers401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListCustomers403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListCustomers404ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListCustomers500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListIntegrations200ResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "name": { "type": "string", "maxLength": 255 },
                "slug": { "type": "string", "maxLength": 255 },
                "description": { "nullable": true, "type": "string" },
                "logo_url": {
                  "nullable": true,
                  "type": "string",
                  "maxLength": 500
                },
                "target_endpoint_url": { "type": "string", "maxLength": 500 },
                "subscribed_events": {
                  "anyOf": [
                    {
                      "anyOf": [
                        { "type": "string" },
                        { "type": "number" },
                        { "type": "boolean" },
                        { "type": "string", "nullable": true, "enum": [null] }
                      ]
                    },
                    { "type": "object", "additionalProperties": {} },
                    { "type": "array", "items": {} }
                  ]
                },
                "config_schema": {
                  "anyOf": [
                    {
                      "anyOf": [
                        { "type": "string" },
                        { "type": "number" },
                        { "type": "boolean" },
                        { "type": "string", "nullable": true, "enum": [null] }
                      ]
                    },
                    { "type": "object", "additionalProperties": {} },
                    { "type": "array", "items": {} }
                  ]
                },
                "is_global": { "type": "boolean" },
                "created_at": {},
                "updated_at": {}
              },
              "required": [
                "id",
                "name",
                "slug",
                "description",
                "logo_url",
                "target_endpoint_url",
                "subscribed_events",
                "config_schema",
                "is_global",
                "created_at",
                "updated_at"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "ListIntegrations401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListIntegrations403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListIntegrations404ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListIntegrations500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key required for authentication."
      }
    }
  },
...
  "security": [{ "x-api-key": [] }]
}